server/requirements.txt
Flask
Flask-Cors
gunicorn
google-cloud-secret-manager
firebase-admin
google-generativeai